// packages/utils/src/enums/cuisine.enum.ts
export enum CuisineType {
  // African Cuisines
  AFRICAN = "african",
  ETHIOPIAN = "ethiopian",
  MOROCCAN = "moroccan",
  NIGERIAN = "nigerian",
  SOUTH_AFRICAN = "south_african",

  // American Cuisines
  AMERICAN = "american",
  BBQ = "bbq",
  SOUTHERN = "southern",
  TEX_MEX = "tex_mex",
  CAJUN = "cajun",

  // Asian Cuisines
  ASIAN = "asian",
  CHINESE = "chinese",
  JAPANESE = "japanese",
  KOREAN = "korean",
  THAI = "thai",
  VIETNAMESE = "vietnamese",
  INDIAN = "indian",
  PAKISTANI = "pakistani",
  FILIPINO = "filipino",
  INDONESIAN = "indonesian",
  MALAYSIAN = "malaysian",

  // European Cuisines
  ITALIAN = "italian",
  FRENCH = "french",
  GERMAN = "german",
  SPANISH = "spanish",
  GREEK = "greek",
  BRITISH = "british",
  IRISH = "irish",
  RUSSIAN = "russian",
  TURKISH = "turkish",
  PORTUGUESE = "portuguese",
  EUROPEAN = "european",

  // Middle Eastern Cuisines
  MIDDLE_EASTERN = "middle_eastern",
  LEBANESE = "lebanese",
  PERSIAN = "persian",
  ISRAELI = "israeli",
  ARAB = "arab",

  // Latin American Cuisines
  MEXICAN = "mexican",
  BRAZILIAN = "brazilian",
  ARGENTINIAN = "argentinian",
  PERUVIAN = "peruvian",
  COLOMBIAN = "colombian",
  CARIBBEAN = "caribbean",

  // Other Categories
  MEDITERRANEAN = "mediterranean",
  SEAFOOD = "seafood",
  VEGETARIAN = "vegetarian",
  VEGAN = "vegan",
  HALAL = "halal",
  KOSHER = "kosher",
  FUSION = "fusion",
  FAST_FOOD = "fast_food",
  STREET_FOOD = "street_food",
  FINE_DINING = "fine_dining",
  CAFE = "cafe",
  BAKERY = "bakery",
  DESSERTS = "desserts",
  HEALTHY = "healthy",
  ORGANIC = "organic",
  GLUTEN_FREE = "gluten_free",
}